refactor: moved images to assets folder
