+++ src/scripts/components/Header/Auth.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import classes from './Auth.module.scss';
-import sprite from '../../../img/header-sprite.svg';
+import sprite from '../../../assets/img/header-sprite.svg';
 import MenuList from '../MenuList/MenuList';
 import MenuListItem from '../MenuList/MenuListItem';
