+++ src/scripts/components/Header/Logo.module.scss
@@ -8,6 +8,6 @@
   div {
     height: 5rem;
     width: 5rem;
-    background: center/cover url('../../../img/logo-72x72.png');
+    background: center/cover url('../../../assets/img/logo-72x72.png');
   }
 }
