+++ src/scripts/components/Table/Stars.module.scss
@@ -8,7 +8,7 @@
 }
 
 .star {
-  background: center/cover url('../../../img/star.svg');
+  background: center/cover url('../../../assets/img/star.svg');
   width: 1.32rem;
   height: 1.56rem;
   margin-right: 0.4rem;
