+++ src/scss/components/_logo.scss
@@ -5,6 +5,6 @@
   div {
     height: 5rem;
     width: 5rem;
-    background: center/cover url('../img/logo-72x72.png');
+    background: center/cover url('../../assets/img/logo-72x72.png');
   }
 }
