+++ src/scss/components/_star.scss
@@ -1,5 +1,5 @@
 .star {
-  background: center/cover url('/src/img/star.svg');
+  background: center/cover url('/src/assets/img/star.svg');
   width: 1.32rem;
   height: 1.56rem;
   margin-right: 0.4rem;
